+++ src/main/java/controller/BlackJackGame.java
@@ -3,10 +3,10 @@ package controller;
 import domain.*;
 import domain.card.Card;
 import domain.deck.Deck;
-import domain.gamer.Dealer;
-import domain.gamer.Gamer;
-import domain.gamer.Name;
-import domain.gamer.Player;
+import domain.gamer.*;
+import dto.DealerResult;
+import dto.PlayerResults;
+import dto.TotalResult;
 import strategy.RandomShuffleStrategy;
 import view.InputView;
 import view.OutputView;
