+++ src/main/java/dto/DealerResult.java
@@ -1,4 +1,6 @@
-package domain;
+package dto;
+
+import domain.Result;
 
 public class DealerResult {
